+++ Contents/Code/siteMylf.py
@@ -12,7 +12,7 @@ def search(results,encodedTitle,title,searchTitle,siteNum,lang,searchByDateActor
         Log("searchString formatted")
     searchResults = HTML.ElementFromURL(PAsearchSites.getSearchSearchURL(siteNum) + searchString)
 
-    titleNoFormatting = searchResults.xpath('//div[@class="col-12 col-md-7"]//span[contains(@class,"text-lightgray")]')[0].text_content().strip()
+    titleNoFormatting = searchResults.xpath('//div[@class="col-12 col-md-8"]//span[contains(@class,"text-lightgray")]')[0].text_content().strip()
     Log("titleNoFormatting: " + titleNoFormatting)
     curID = searchResults.xpath('//link[@rel="canonical"]')[0].get('href').replace('/','_').replace('?','!')
     Log("curID: " + curID)
@@ -20,7 +20,10 @@ def search(results,encodedTitle,title,searchTitle,siteNum,lang,searchByDateActor
     Log("# actors: " + str(len(actors)))
     firstActor = actors[0].text_content()
     Log("firstActor: " + firstActor)
-    subSite = searchResults.xpath('//img[@class="lazy img-fluid"]')[0].get("data-original").split('/')[-1].replace('_logo.png','').title()
+    if "mylfdom" in curID:
+        subSite = "MylfDom"
+    else:
+        subSite = searchResults.xpath('//img[@class="lazy img-fluid"]')[0].get("data-original").split('/')[-1].replace('_logo.png','').title()
     Log("subSite: " + subSite)
     if searchDate:
         releaseDate = parse(searchDate).strftime('%Y-%m-%d')
@@ -36,11 +39,11 @@ def update(metadata,siteID,movieGenres,movieActors):
 
     url = str(metadata.id).split("|")[0].replace('_','/').replace('?','!')
     detailsPageElements = HTML.ElementFromURL(url)
-    # art = []
+    art = []
     metadata.collections.clear()
     movieGenres.clearGenres()
     movieActors.clearActors()
-    urlBase = PAsearchSites.getSearchBaseURL(siteID)
+
     # Studio
     metadata.studio = 'Mylf'
 
@@ -48,16 +51,15 @@ def update(metadata,siteID,movieGenres,movieActors):
     metadata.title = detailsPageElements.xpath('//span[contains(@class,"m_scenetitle")]')[0].text_content().strip()
     Log("title: " + metadata.title)
 
+    # Summary
+    summary = detailsPageElements.xpath('//div[contains(@class,"text-light")]')[0].text_content().strip()
+    metadata.summary = summary.replace('See full video here >', '')
+
     #Tagline and Collection(s)
     subSite = PAsearchSites.getSearchSiteName(siteID)
     metadata.tagline = subSite
     metadata.collections.add(subSite)
     Log("subSite: " + subSite)
-
-    # Summary
-    summary = detailsPageElements.xpath('//div[contains(@class,"text-light")]')[0].text_content().strip()
-    metadata.summary = summary.replace('See full video here >', '')
-
 
     # Release Date
     date = str(metadata.id).split("|")[2]
@@ -72,31 +74,13 @@ def update(metadata,siteID,movieGenres,movieActors):
     if len(actors) > 0:
         for actor in actors:
             actorName = actor.xpath('./span')[0].text_content().strip()
-            actorPageURL = urlBase + actor.get("href")
+            actorPageURL = actor.get("href")
+            if 'http' not in actorPageURL:
+                actorPageURL = PAsearchSites.getSearchBaseURL(siteID) + actorPageURL
             actorPage = HTML.ElementFromURL(actorPageURL)
             actorPhotoURL = actorPage.xpath('//img[contains(@class,"girlthumb")]')[0].get("data-original")
             movieActors.addActor(actorName,actorPhotoURL)
             Log('actor: ' + actorName + ", " + actorPhotoURL)
-
-    ### Posters and artwork ###
-    posterNum = 1
-    posters = detailsPageElements.xpath('//div[contains(@class,"trailer-small-images")]//a')
-    for poster in posters:
-        posterURL = poster.get("href")
-        if posterURL != "/join":
-            if 'http' not in posterURL:
-                posterURL = urlBase + posterURL
-            metadata.posters[posterURL] = Proxy.Preview(HTTP.Request(posterURL, headers={'Referer': 'http://www.google.com'}).content, sort_order = posterNum)
-            metadata.art[posterURL] = Proxy.Preview(HTTP.Request(posterURL, headers={'Referer': 'http://www.google.com'}).content, sort_order = posterNum)
-            posterNum += 1
-            Log('posterURL: ' + posterURL)
-
-    # Video trailer background image
-    previewBG = detailsPageElements.xpath('//video[@id="main-movie-player"]')[0].get('poster')
-    if 'http' not in previewBG:
-        previewBG = urlBase + previewBG
-    metadata.art[previewBG] = Proxy.Preview(HTTP.Request(previewBG, headers={'Referer': 'http://www.google.com'}).content, sort_order = posterNum)
-    Log('previewIMG: ' + previewBG)
 
     # Genres
     movieGenres.clearGenres()
@@ -129,9 +113,51 @@ def update(metadata,siteID,movieGenres,movieActors):
     elif subSite.lower() == "Mylfed".lower():
         for genreName in ['Lesbian', 'Girl on Girl', 'GG']:
             movieGenres.addGenre(genreName)
+    elif subSite.lower() == "MylfDom".lower():
+        for genreName in ['BDSM']:
+            movieGenres.addGenre(genreName)
     if (len(actors) > 1) and subSite != "Mylfed":
         genres.append("Threesome")
     for genre in genres:
         movieGenres.addGenre(genre)
+
+    ### Posters and artwork ###
+
+    # Video trailer background image
+    try:
+        twitterBG = detailsPageElements.xpath('//video[@id="main-movie-player"]')[0].get('poster')
+        art.append(twitterBG)
+    except:
+        pass
+
+    # Photos
+    photos = detailsPageElements.xpath('//div[contains(@class,"trailer-small-images")]//a')
+    if len(photos) > 0:
+        for photoLink in photos:
+            photo = photoLink.get('href')
+            if 'http' not in photo:
+                photo = PAsearchSites.getSearchBaseURL(siteID) + photo
+            art.append(photo)
+
+    j = 1
+    Log("Artwork found: " + str(len(art)))
+    for posterUrl in art:
+        if not PAsearchSites.posterAlreadyExists(posterUrl,metadata):
+            #Download image file for analysis
+            try:
+                img_file = urllib.urlopen(posterUrl)
+                im = StringIO(img_file.read())
+                resized_image = Image.open(im)
+                width, height = resized_image.size
+                #Add the image proxy items to the collection
+                if width > 1 or height > width:
+                    # Item is a poster
+                    metadata.posters[posterUrl] = Proxy.Preview(HTTP.Request(posterUrl, headers={'Referer': 'http://www.google.com'}).content, sort_order = j)
+                if width > 100 and width > height:
+                    # Item is an art item
+                    metadata.art[posterUrl] = Proxy.Preview(HTTP.Request(posterUrl, headers={'Referer': 'http://www.google.com'}).content, sort_order = j)
+                j = j + 1
+            except:
+                pass
 
     return metadata